requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.9.2
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
coverage==7.5.3
memory-profiler==0.61.0
sphinx==7.3.7
sphinx-rtd-theme==2.0.0

